Create different ChargePoint based on subprotocol
This also changes default subprotocol setting from OCPP 1.6 to 1.6 and
2.0.1.
