+++ custom_components/ocpp/const.py
@@ -49,7 +49,8 @@ DEFAULT_FORCE_SMART_CHARGING = False
 DEFAULT_SSL = False
 DEFAULT_SSL_CERTFILE_PATH = pathlib.Path.cwd().joinpath("fullchain.pem")
 DEFAULT_SSL_KEYFILE_PATH = pathlib.Path.cwd().joinpath("privkey.pem")
-DEFAULT_SUBPROTOCOL = "ocpp1.6"
+DEFAULT_SUBPROTOCOL = "ocpp1.6,ocpp2.0.1"
+OCPP_2_0 = "ocpp2.0"
 DEFAULT_METER_INTERVAL = 60
 DEFAULT_IDLE_INTERVAL = 900
 DEFAULT_WEBSOCKET_CLOSE_TIMEOUT = 10
